gcin-0.7.3

+++ im-client/gcin-crypt.c
@@ -1,13 +1,13 @@
 #include "gcin-protocol.h"
 
-static int __gcin_rand__(u_long *next)
+static int __gcin_rand__(u_int *next)
 {
   *next = *next * 1103515245 + 12345;
   return((unsigned)(*next/65536) % 32768);
 }
 
 void __gcin_enc_mem(u_char *p, int n,
-                    GCIN_PASSWD *passwd, u_long *seed)
+                    GCIN_PASSWD *passwd, u_int *seed)
 {
   int i;
